pyproject.toml
[project]
name = "ppt2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.2.5",
    "opencv-python>=4.11.0.86",
    "paddleocr>=2.10.0",
    "paddlepaddle>=3.0.0",
    "pillow>=11.2.1",
    "psutil>=7.0.0",
    "python-pptx>=1.0.2",
    "requests>=2.32.3",
    "setuptools>=80.4.0",
]

[dependency-groups]
dev = [
    "huggingface-hub>=0.31.2",
]
